backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.2.0
gunicorn==22.0.0
whitenoise==6.6.0
redis==5.0.4
django-redis==5.4.0
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
requests==2.32.2
beautifulsoup4==4.12.3
lxml==5.2.2
pypdf==4.2.0
pdfplumber==0.11.0
python-dateutil==2.9.0
pytz==2024.1
Pillow==10.3.0
drf-spectacular==0.27.2
django-ratelimit==4.1.0
argon2-cffi==23.1.0
cryptography==42.0.7
sentry-sdk==2.3.1
